cleanup app for routes
